+++ tests/python/test_stiffness_checker.py
@@ -12,18 +12,13 @@ from numpy.testing import assert_equal, assert_almost_equal
 
 from pyconmech import StiffnessChecker
 from pyconmech.frame_analysis import read_frame_json, read_load_case_json, check_material_dict
-# from pyconmech.database import MATERIAL_PROPERTY_DATABASE
-
-# def test_get_material_from_database():
-#     PLA_MAT = MATERIAL_PROPERTY_DATABASE['PLA']
-#     STEEL_S235_MAT = MATERIAL_PROPERTY_DATABASE['Steel-S235']
-#     assert check_material_dict(PLA_MAT)
-#     assert check_material_dict(STEEL_S235_MAT)
+from pyconmech.frame_analysis import PointLoad, GravityLoad, UniformlyDistLoad
 
 def repetitive_test_stiffness_checker(frame_file_path, engine, parsed_pt_loads=None, include_sw=False,
     n_attempts=50, existing_ids=[], expect_partial_ids_success=True, trans_tol=2e-3):
-    sc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine=engine)
-    sc.set_loads(point_loads=parsed_pt_loads, include_self_weight=include_sw)
+    sc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine=engine, verbose=True)
+    gravity_load = GravityLoad([0,0,-1]) if include_sw else None
+    sc.set_loads(point_loads=parsed_pt_loads, gravity_load=gravity_load)
     assert not sc.has_stored_result()
     sc.set_nodal_displacement_tol(trans_tol=trans_tol)
 
@@ -31,10 +26,11 @@ def repetitive_test_stiffness_checker(frame_file_path, engine, parsed_pt_loads=N
 
     trans_tol, rot_tol = sc.get_nodal_deformation_tol()
     max_trans, max_rot, max_t_id, max_r_id = sc.get_max_nodal_deformation()
-    print('sol benchmark: max_trans {0} / tol {4}, max_rot {1} / tol {5}, max_trans_id {2}, max_rot_id {3}'.format(\
-        max_trans, max_rot, max_t_id, max_r_id, trans_tol, rot_tol))
+    print('sol benchmark: success? {6} | max_trans {0} / tol {4}, max_rot {1} / tol {5}, max_trans_id {2}, max_rot_id {3}'.format(\
+        max_trans, max_rot, max_t_id, max_r_id, trans_tol, rot_tol, sol_success))
 
     if expect_partial_ids_success:
+        assert sc.get_compliance() > 0.0
         assert sol_success, 'max_trans {}, max_rot {}, max_trans_id {}, max_rot_id {}'.format(*sc.get_max_nodal_deformation())
     else:
         assert not sol_success, 'max_trans {}, max_rot {}, max_trans_id {}, max_rot_id {}'.format(*sc.get_max_nodal_deformation())
@@ -77,7 +73,7 @@ def repetitive_test_stiffness_checker(frame_file_path, engine, parsed_pt_loads=N
     st_time = time.time()
     for _ in range(n_attempts):
         sc_new = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine=engine)
-        sc_new.set_loads(point_loads=parsed_pt_loads, include_self_weight=include_sw)
+        sc_new.set_loads(point_loads=parsed_pt_loads, gravity_load=gravity_load)
         sc_new.set_nodal_displacement_tol(trans_tol=trans_tol)
 
         shuffled_existing_ids = copy(existing_ids)
@@ -95,7 +91,8 @@ def repetitive_test_stiffness_checker(frame_file_path, engine, parsed_pt_loads=N
             assert_almost_equal(eR[i][0], tmp_eR[i][0]) 
             assert_almost_equal(eR[i][1], tmp_eR[i][1]) 
     tot_time = time.time() - st_time
-    print('\nwith reinit: avg time: {} s, total_time: {} s for {} runs'.format(tot_time / n_attempts, tot_time, n_attempts))
+    print('with reinit: avg time: {} s, total_time: {} s for {} runs'.format(tot_time / n_attempts, tot_time, n_attempts))
+    cprint('Pass.', 'green')
 
 
 @pytest.mark.rep_test_sc
@@ -105,7 +102,7 @@ def repetitive_test_stiffness_checker(frame_file_path, engine, parsed_pt_loads=N
 def test_stiffness_checker_solve_consistency(test_case, load_case, test_data_dir, n_attempts, engine):
     if test_case == 'tower':
         file_name = 'tower_3D.json'
-        load_file_name = 'tower_3D_load_case.json'
+        load_file_name = 'tower_3D_loads.json'
         
         # ! some parts are floating in this partial ids
         failed_existing_ids = [0, 4, 7, 8, 9] 
@@ -114,13 +111,13 @@ def test_stiffness_checker_solve_consistency(test_case, load_case, test_data_dir
             trans_tol = 0.04
         else:
             success_existing_ids = [0,1,2,3,8,9,10,11,12,13,14,15]
-            trans_tol = 0.001
+            trans_tol = 0.0015
     elif test_case == 'topopt-100':
         file_name = 'topopt-100.json'
-        load_file_name = 'topopt-100_load_case.json'
+        load_file_name = 'topopt-100_loads.json'
 
         # a big cantilever are floating in this partial ids
-        failed_existing_ids = [125, 126, 115, 122, 111, 108, 23, 22, 98, 75, 64, 34, 61, 65, 59, 60, 39, 36, 44, 67]
+        failed_existing_ids = [1, 56, 59, 60, 66, 67, 68, 69, 83, 84, 102, 130]
 
         success_existing_ids = [0, 1, 2, 3, 4, 5, 6, 8, 11, 14, 15, 19, 20, 21, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33,
                 34, 35, 36, 37, 38, 41, 42, 43, 44, 46, 51, 52, 53, 54, 55, 76, 88, 89, 90, 91, 92,
@@ -156,55 +153,26 @@ def test_stiffness_checker_solve_consistency(test_case, load_case, test_data_dir
     repetitive_test_stiffness_checker(json_path, engine, parsed_pt_loads=parsed_pt_loads, include_sw=include_sw, \
         n_attempts=n_attempts, existing_ids=failed_existing_ids, expect_partial_ids_success=False, trans_tol=trans_tol)
 
-
-@pytest.mark.init_sc
-def test_init_stiffness_checker(test_data_dir, engine):
-    # file_name = 'tower_3D_broken_lines.json'
-    file_name = 'four-frame.json'
-    file_path = os.path.join(test_data_dir, file_name)
-
-    print('init StiffnessChecker from json file path...')
-    sc_from_json = StiffnessChecker.from_json(file_path, verbose=False, checker_engine=engine)
-
-    print('init StiffnessChecker from frame data...')
-    nodes, elements, fix_specs, model_type, material_dicts, model_name, unit = \
-        read_frame_json(file_path, verbose=True)
-    sc_from_data = StiffnessChecker.from_frame_data(nodes, elements, fix_specs, material_dicts, 
-        unit=unit, model_type=model_type, model_name=model_name, verbose=False, checker_engine=engine)
-
-    assert sc_from_json.model_name == sc_from_data.model_name
-    assert sc_from_json.model_type == sc_from_data.model_type
-    for mat1, mat2 in zip(sc_from_json.materials, sc_from_json.materials):
-        assert mat1 == mat2
-    for n1, n2 in zip(sc_from_json.node_points, sc_from_data.node_points):
-        assert_equal(n1, n2)
-    for e1, e2 in zip(sc_from_json.elements, sc_from_data.elements):
-        assert_equal(e1, e2)
-    for fv1, fv2 in zip(sc_from_json.fix_node_ids, sc_from_data.fix_node_ids):
-        assert_equal(fv1, fv2)
-    for vid, spec in sc_from_json.fix_specs.items():
-        assert vid in sc_from_data.fix_specs
-        assert_equal(spec, sc_from_data.fix_specs[vid])
-
-
 @pytest.mark.nodal_load
 def test_get_nodal_loads(engine):
     file_name = 'tower_3D.json'
-    load_file_name = 'tower_3D_load_case.json'
+    load_file_name = 'tower_3D_loads.json'
     root_dir = os.path.dirname(os.path.abspath(__file__))
     json_path = os.path.join(root_dir, '..', 'test_data', file_name)
     load_json_path = os.path.join(root_dir, '..', 'test_data', load_file_name)
 
     sc = StiffnessChecker.from_json(json_file_path=json_path, checker_engine=engine)
-    sc.set_self_weight_load(False)
+    sc.set_self_weight_load(None)
     parsed_pt_loads, _, _ = read_load_case_json(load_json_path)
 
     sc.set_loads(point_loads=parsed_pt_loads)
     fetched_nodal_loads = sc.get_nodal_loads()
 
-    for node_id in parsed_pt_loads:
+    for pt_load in parsed_pt_loads:
+        assert isinstance(pt_load, PointLoad)
+        node_id = pt_load.node_ind
         assert node_id in fetched_nodal_loads
-        assert_equal(parsed_pt_loads[node_id], fetched_nodal_loads[node_id])
+        assert_equal(pt_load.force + pt_load.moment, fetched_nodal_loads[node_id])
 
 
 @pytest.mark.neighnor_q
@@ -215,34 +183,37 @@ def test_neighnor_query(engine):
     json_path = os.path.join(root_dir, '..', 'test_data', file_name)
 
     sc = StiffnessChecker.from_json(json_file_path=json_path, checker_engine=engine)
-    assert_equal(sc.fix_element_ids, [0,1,2,3,8,9,10,11,12,13,14,15])
-    assert_equal(sc.get_element_connected_node_ids(fix_node_only=True), [0,1,2,3]) 
-    assert_equal(sc.get_element_connected_node_ids(existing_ids=[0,3], fix_node_only=True), [1,2]) 
+    assert_equal(sc.fix_element_ids, [0, 1, 2, 3, 8, 9, 10, 11, 12, 13, 14, 15])
+    assert_equal(sc.get_element_connected_node_ids(fix_node_only=True), [0,2,4,6]) 
+    assert_equal(sc.get_element_connected_node_ids(existing_ids=[0,3], fix_node_only=True), [0,6]) 
+    assert_equal(set(sc.get_element_connected_node_ids(existing_ids=[0,3], fix_node_only=False)), set([0,6,1,7])) 
 
 
 def repetitive_test_equilibrium(frame_file_path, engine, parsed_pt_loads=None, include_sw=False,
         existing_ids=[], expect_partial_ids_success=True, debug=False, trans_tol=1e-3):
     sc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine=engine, verbose=True)
     sc.set_nodal_displacement_tol(trans_tol=trans_tol)
-    sc.set_loads(point_loads=parsed_pt_loads, include_self_weight=include_sw)
+    gravity_load = GravityLoad([0,0,-1.]) if include_sw else None
+    sc.set_loads(point_loads=parsed_pt_loads, gravity_load=gravity_load)
     assert not sc.has_stored_result()
 
     sol_success = sc.solve(existing_ids, eid_sanity_check=False, if_cond_num=True)
-    if debug:
+    if debug and sol_success:
         trans_tol, _ = sc.get_nodal_deformation_tol()
         max_t, _, max_tid, _ = sc.get_max_nodal_deformation()
         cprint('{} | Max trans deformation: Node #{}: {}|{}'.format(engine, max_tid, max_t, trans_tol), 'cyan')
 
-        scc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine='cpp', verbose=False)
-        scc.set_nodal_displacement_tol(trans_tol=trans_tol)
-        scc.set_loads(point_loads=parsed_pt_loads, include_self_weight=include_sw)
+        # scc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine='cpp', verbose=False)
+        # scc.set_nodal_displacement_tol(trans_tol=trans_tol)
+        # scc.set_loads(point_loads=parsed_pt_loads, gravity_load=gravity_load)
 
-        scc_success = scc.solve(existing_ids, eid_sanity_check=False, if_cond_num=True)
-        scc_trans_tol, _ = scc.get_nodal_deformation_tol()
-        scc_max_t, _, scc_max_tid, _ = scc.get_max_nodal_deformation()
-        cprint('cpp | Max trans deformation: Node #{}: {}|{}'.format(scc_max_tid, scc_max_t, scc_trans_tol), 'magenta')
+        # scc_success = scc.solve(existing_ids, eid_sanity_check=False, if_cond_num=True)
+        # scc_trans_tol, _ = scc.get_nodal_deformation_tol()
+        # scc_max_t, _, scc_max_tid, _ = scc.get_max_nodal_deformation()
+        # cprint('cpp | Max trans deformation: Node #{}: {}|{}'.format(scc_max_tid, scc_max_t, scc_trans_tol), 'magenta')
 
     if expect_partial_ids_success:
+        assert sc.get_compliance() > 0
         assert sol_success
     else:
         assert not sol_success
@@ -400,21 +371,22 @@ def test_nodal_equilibrium(test_case, load_case, engine, debug):
     trans_tol = 1e-3
     if test_case == 'tower':
         file_name = 'tower_3D.json'
-        load_file_name = 'tower_3D_load_case.json'
+        load_file_name = 'tower_3D_loads.json'
         
         # some parts are floating in this partial ids
-        failed_existing_ids = [0, 4, 7, 8, 9] 
+        failed_existing_ids = [0, 5, 6, 12, 15, 16]        
         if load_case == 'self_weight' or load_case == 'self_weight+point_load':
             trans_tol = 0.23
             success_existing_ids = [0, 1, 2, 3, 4, 5, 6, 10, 13]
         else:
+            trans_tol = 0.0015
             success_existing_ids = [0,1,2,3,8,9,10,11,12,13,14,15]
     elif test_case == 'topopt-100':
         file_name = 'topopt-100.json'
-        load_file_name = 'topopt-100_load_case.json'
+        load_file_name = 'topopt-100_loads.json'
 
         # a big cantilever are floating in this partial ids
-        failed_existing_ids = [125, 126, 115, 122, 111, 108, 23, 22, 98, 75, 64, 34, 61, 65, 59, 60, 39, 36, 44, 67]
+        failed_existing_ids = [1, 56, 59, 60, 66, 67, 68, 69, 83, 84, 102, 130]
         success_existing_ids = [0, 1, 2, 3, 4, 5, 8, 11, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, \
         43, 44, 51, 52, 53, 88, 89, 90, 91, 92, 93, 94, 96, 97, 98, 102, 103, 104, 105, 106, 107, \
         108, 109, 110, 111, 112, 113, 114, 115, 117, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131]
@@ -453,12 +425,12 @@ def repetitive_check_gravity_validity(frame_file_path, engine, existing_ids=[], 
     sc = StiffnessChecker.from_json(json_file_path=frame_file_path, checker_engine=engine)
 
     print('\n################\n [0,0,-100] gravity direction')
-    sc.set_loads(include_self_weight=True, gravity_direction=[0,0,-100])
+    sc.set_loads(gravity_load=GravityLoad([0,0,-100]))
     sc.solve(existing_ids, eid_sanity_check=True)
     compliance_big = sc.get_compliance()
 
     print('\n################\n [0,0,-1] gravity direction')
-    sc.set_loads(include_self_weight=True, gravity_direction=[0,0,-1])
+    sc.set_loads(gravity_load=GravityLoad([0,0,-1]))
     sc.solve(existing_ids, eid_sanity_check=True)
     compliance_small = sc.get_compliance()
 
@@ -479,21 +451,20 @@ def repetitive_check_gravity_validity(frame_file_path, engine, existing_ids=[], 
     # all_node_e_neighbors = sc.get_node_neighbors(return_e_id=True)
 
     # return node_points, element_vids, fix_specs, model_type, material_dicts, model_name, unit
-    node_points, element_vids, _, _, material_dicts, _, unit = \
-        read_frame_json(frame_file_path)
+    node_points, element_vids, _, _, _, _, _, unit = read_frame_json(frame_file_path)
 
     # direct total gravity calculation
     total_weight_force = 0
     for e_id in existing_e_ids:
         n1, n2 = sc.elements[e_id]
         # convert to meter
-        e_len = norm(np.array(node_points[n1]) - np.array(node_points[n2]))
+        e_len = norm(np.array(node_points[n1].point) - np.array(node_points[n2].point))
         if unit != 'meter':
             e_len *= 1e-3
-        assert material_dicts[e_id]['cross_sec_area_unit'] == "centimeter^2" and material_dicts[e_id]["density_unit"] == "kN/m3", \
-            'E#{}: {}'.format(e_id, material_dicts[e_id])
-        rho = material_dicts[e_id]['density'] 
-        A = material_dicts[e_id]['cross_sec_area'] * 1e-4 # convert to meter^2
+        # assert material_dicts[e_id]['cross_sec_area_unit'] == "centimeter^2" and material_dicts[e_id]["density_unit"] == "kN/m3", \
+        #     'E#{}: {}'.format(e_id, material_dicts[e_id])
+        rho = sc.get_element_material(e_id).density
+        A = sc.get_element_crosssec(e_id).A # meter^2
         total_weight_force += e_len * A * rho
     
     # total gravity from lumped nodal loads
@@ -545,10 +516,12 @@ def test_self_weight_validity(test_case, engine):
 @pytest.mark.uniform_load_check
 @pytest.mark.parametrize("test_case, existing_e_ids", 
     [('tower', []), ('tower', list(range(8))), 
-     ('topopt-100', []), ('topopt-100', [0, 1, 2, 3, 4, 5, 8, 11, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, \
-        43, 44, 51, 52, 53, 88, 89, 90, 91, 92, 93, 94, 96, 97, 98, 102, 103, 104, 105, 106, 107, \
-        108, 109, 110, 111, 112, 113, 114, 115, 117, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131])
+     ('topopt-100', [])
      ])
+    #  ('topopt-100', [0, 1, 2, 3, 4, 5, 8, 11, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, \
+    #     43, 44, 51, 52, 53, 88, 89, 90, 91, 92, 93, 94, 96, 97, 98, 102, 103, 104, 105, 106, 107, \
+    #     108, 109, 110, 111, 112, 113, 114, 115, 117, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131])
+    #  ])
 def test_uniformly_distributed_load_with_gravity(test_case, existing_e_ids, engine):
     if test_case == 'tower':
         file_name = 'tower_3D.json'
@@ -563,18 +536,17 @@ def test_uniformly_distributed_load_with_gravity(test_case, existing_e_ids, engi
 
     # gravity load agreement check
     # _, uniform_element_load, _ = read_load_case_json(load_json_path)
-    node_points, element_vids, _, _, material_dicts, _, _ = \
-        read_frame_json(json_path)
+    node_points, element_vids, _, _, _, _, _, unit = read_frame_json(json_path)
 
-    uniform_distributed_load = {}
+    uniform_distributed_load = []
     for e_id in range(len(element_vids)):
-        assert material_dicts[e_id]['cross_sec_area_unit'] == 'centimeter^2' and material_dicts[e_id]["density_unit"] == "kN/m3"
-        A = material_dicts[e_id]['cross_sec_area'] * 1e-4 # convert to meter^2
-        rho = material_dicts[e_id]['density']
-        uniform_distributed_load[e_id] = [0, 0, - rho * A]
+        # assert material_dicts[e_id]['cross_sec_area_unit'] == 'centimeter^2' and material_dicts[e_id]["density_unit"] == "kN/m3"
+        rho = sc.get_element_material(e_id).density
+        A = sc.get_element_crosssec(e_id).A # meter^2
+        uniform_distributed_load.append(UniformlyDistLoad([0, 0, - rho * A], None, [""]))
 
     # existing_e_ids = list(range(len(sc.elements)))
-    sc.set_loads(uniform_distributed_load=uniform_distributed_load, include_self_weight=False)
+    sc.set_loads(uniform_distributed_load=uniform_distributed_load)
 
     nodal_loads = sc.get_nodal_loads(existing_ids=existing_e_ids)
     sw_loads = sc.get_self_weight_loads(existing_ids=existing_e_ids)
@@ -585,11 +557,11 @@ def test_uniformly_distributed_load_with_gravity(test_case, existing_e_ids, engi
     _, ul_nD, ul_fR, ul_eR = sc.get_solved_results()
 
     sw_sc = StiffnessChecker.from_json(json_file_path=json_path, checker_engine=engine)
-    sw_sc.set_loads(include_self_weight=True)
+    sw_sc.set_loads(gravity_load=GravityLoad([0,0,-1]))
     sw_sc.solve(exist_element_ids=existing_e_ids)
     _, sw_nD, sw_fR, sw_eR = sw_sc.get_solved_results()
 
-    # checking explicit, manually constructed gravity load consistent with the result from setting include_self_weight
+    # checking explicit, manually constructed gravity load consistent with the result from setting gravity_load
     for n_id in ul_nD.keys():
         assert_almost_equal(ul_nD[n_id], sw_nD[n_id])
     for f_id in ul_fR.keys():
@@ -611,21 +583,22 @@ def test_uniformly_distributed_load_with_analytical_solution(engine, n_attempts)
             Page 108, section 5.2 Loads between nodal points
     """
     file_name = 'uniform_load_analytical_model.json'
-    load_file_name = 'uniform_load_analytical_model_load_case.json'
+    load_file_name = 'uniform_load_analytical_model_loads.json'
 
     root_dir = os.path.dirname(os.path.abspath(__file__))
     json_path = os.path.join(root_dir, '..', 'test_data', file_name)
     load_json_path = os.path.join(root_dir, '..', 'test_data', load_file_name)
 
     sc = StiffnessChecker.from_json(json_file_path=json_path, checker_engine=engine)
-    point_load, uniform_element_load, include_sw = read_load_case_json(load_json_path)
-    sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, include_self_weight=include_sw)
+    point_load, uniform_element_load, gravity_load = read_load_case_json(load_json_path)
+    print(uniform_element_load)
+    sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, gravity_load=gravity_load)
     sc.set_nodal_displacement_tol(trans_tol=0.024, rot_tol=0.006)
 
     def compare_analytical_sol(pass_criteria, nD, fR, eR, nodal_loads, check_decimal=1):
-        assert_equal(nodal_loads[1][2], -12.5)
-        assert_equal(nodal_loads[1][3], 0.0)
-        assert_equal(nodal_loads[1][4], -6.250)
+        # assert_equal(nodal_loads[1][2], -12.5)
+        # assert_equal(nodal_loads[1][3], 0.0)
+        # assert_equal(nodal_loads[1][4], -6.250)
         # print('{} \t {}'.format(fR[2][0:3], fR[2][3:6]))
 
         assert pass_criteria
@@ -645,8 +618,8 @@ def test_uniformly_distributed_load_with_analytical_solution(engine, n_attempts)
     print('compare analytical res: w reinit')
     for _ in range(n_attempts):
         re_sc = StiffnessChecker.from_json(json_file_path=json_path, checker_engine=engine)
-        point_load, uniform_element_load, include_sw = read_load_case_json(load_json_path)
-        re_sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, include_self_weight=include_sw)
+        point_load, uniform_element_load, gravity_load = read_load_case_json(load_json_path)
+        re_sc.set_loads(point_loads=point_load, uniform_distributed_load=uniform_element_load, gravity_load=gravity_load)
         re_sc.set_nodal_displacement_tol(trans_tol=0.024, rot_tol=0.006)
 
         re_sc.solve()